bear witness to the great migration! basically, added services and helpers to the application layer, moved models to domain
