+++ Treehoot/Treehoot.Domain/Models/Question.cs
@@ -5,5 +5,5 @@ public class Question
     public int Id { get; set; }
     
     public string QuestionText{ get; set; }
-    public List<Answer> Answers { get; set; }
+    public List<int> Answers { get; set; }
 }
