+++ Treehoot/Treehoot.Domain/Models/Quiz.cs
@@ -7,5 +7,5 @@ public class Quiz
     public string Name { get; set; }
     public string Description { get; set; }
     public int Timer { get; set; }
-    public List<Stage> Stages { get; set; }
+    public List<int> Stages { get; set; }
 }
