+++ Treehoot/Treehoot.Domain/Models/Stage.cs
@@ -5,6 +5,6 @@ public class Stage
     public int Id { get; set; }
 
     public string Name { get; set; }
-    public List<Question> Questions { get; set; }
+    public List<int> Questions { get; set; }
 }
