+++ Treehoot/Treehoot.Domain/Models/StageFull.cs
@@ -1,4 +1,4 @@
-﻿namespace Treehoot_API.Models;
+﻿namespace Treehoot.Domain.Models;
 
 public class StageFull
 {
