+++ Treehoot/Treehoot.Domain/Models/User.cs
@@ -1,9 +1,11 @@
-﻿namespace Treehoot.Domain.Models;
+﻿using System.ComponentModel.DataAnnotations;
 
+namespace Treehoot.Domain.Models;
 public class User
 {
     public Guid Id { get; set; }
     public string Name { get; set; }
     public string Email { get; set; }
+
 }
