--- Treehoot/Treehoot_API/Models/Answer.cs
@@ -1,9 +1,0 @@
-﻿namespace Treehoot_API.Models;
-
-public class Answer
-{
-    public int Id { get; set; }
-
-    public bool IsCorrect { get; set; }
-    public string Text { get; set; }
-}
