--- Treehoot/Treehoot_API/Models/Quiz.cs
@@ -1,11 +1,0 @@
-﻿namespace Treehoot_API.Models;
-
-public class Quiz
-{
-    public int Id { get; set; }
-    
-    public string Name { get; set; }
-    public string Description { get; set; }
-    public int Timer { get; set; }
-    public List<int> Stages { get; set; }
-}
