--- Treehoot/Treehoot_API/Models/Stage.cs
@@ -1,10 +1,0 @@
-﻿namespace Treehoot_API.Models;
-
-public class Stage
-{
-    public int Id { get; set; }
-
-    public string Name { get; set; }
-    public List<int> Questions { get; set; }
-}
-
