--- Treehoot/Treehoot_API/Models/User.cs
@@ -1,11 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Treehoot_API.Models;
-public class User
-{
-    public Guid Id { get; set; }
-    public string Name { get; set; }
-    public string Email { get; set; }
-
-}
-
