+++ Treehoot/Treehoot_API/Treehoot_API.csproj
@@ -11,6 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Treehoot.Application\Treehoot.Application.csproj" />
     <ProjectReference Include="..\Treehoot.Domain\Treehoot.Domain.csproj" />
   </ItemGroup>
